configure: respect libdir at build time
The xmlbird.pc file encodes the libdir in it, so make sure we have it
at configure & build time.  That means moving the flag from install to
configure and the config module.

Also clean up the code a bit so `--libdir lib64` doesn't break things.
It previously assumed that the argument always had a leading slash.